requirements.txt
python-decouple
telethon
urlextract